+++ restAudiPro/src/main/java/com/restAudiPro/rest/service/SecuenciaBDService.java
@@ -27,7 +27,7 @@ public class SecuenciaBDService {
         secuenciaBD counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
                 new Update().inc("seq",1), options().returnNew(true).upsert(true),
                 secuenciaBD.class);
-        return (int) (!Objects.isNull(counter) ? counter.getSeq() : 1);
+        return !Objects.isNull(counter) ? counter.getSeq() : 1;
 
     }
 }
